requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
mysql-connector-python==8.3.0
Flask-CORS==4.0.0
